requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
asgiref==3.8.1
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.4
django-ipware==7.0.1
django-jazzmin==3.0.0
django-otp==1.5.0
django-ratelimit==4.1.0
frozenlist==1.4.1
geoip2==4.8.0
idna==3.7
maxminddb==2.6.1
multidict==6.0.5
Pillow==8.2.0
pypng==0.20220715.0
python-ipware==3.0.0
pytz==2021.1
qrcode==7.4.2
requests==2.31.0
setuptools==69.5.1
sqlparse==0.4.1
typing_extensions==4.11.0
urllib3==2.2.1
wheel==0.43.0
yarl==1.9.4
